tests: execute remote container tests
This is mainly for jenkins CI and it is only the first step
(turning on PostgreSQL tests for now)!  In future we should have
some local-only (and less expensive) tests, too.

Jenkins is only responsible for setting 'OS' variable (defaults to
centos7).

When some changes are needed in remote git repositories (say
PostgreSQL container) one have to first submit a pull request
against that git repository where one of the commits needs to
contain commit message with line of this format:

    Required-by: postgresql-container#182

.. where 'postgresql-container' is the name of project within
sclorg group, and '182' string is the id of the already submitted
pull request.  Local testing against such pull-requests works too
(unless the actual branch is not yet merged into 'origin/master'
branch).  There might be arbitrary number of such 'Required-by'
statements.

Fixes: #14