squash: don't install docker-squash with easy_install
That doesn't work unless we use Python version from collection;
so we now just test that the explicit version of docker-squash
tool exists and is in $PATH.  It is up to the end user to install
the build/test requirements anyway, including the docker-squash
utility.

In other words, the commit 0f93964ea15b did not help.