Allow to change the docker build context
In order to allow sharing files in the parent directory, instead of having copies
of the same files for every version, we need to be able to set the context of the
docker build command for every image specifically. This change defines a new
variables DOCKER_BUILD_CONTEXT that may be defined in the Makefile, typically
to '..', so docker build will be run with the context of the parent directory.