Pick oc version by setting PATH
Picking version of OpenShift is not enough to be done by setting --version argument, ideally there should be oc version of the same version. This change first tries to pick oc from /usr/local where different preinstalled versions of oc may already be prepared, and if this is not found, a community released oc binary is downloaded into /tmp.