build: use docker_squash lib, not docker-squash script
The docker-squash script used to print the squashed <image id> on
stderr, but recently it was changed so the image id is printed to
stdout.  To be able to parse image id from both docker_squash
versions (before and after this change), we'd have to have some
ugly if-fork to detect the docker_squash version (or play with FD
redirection).

So rather let's do as upstream suggested in [1] and use the
docker_squash library directly.

Also add a simple CI test for the ./squash.py file.

[1] https://github.com/goldmann/docker-squash/pull/145

Fixes: #101