Make docker build not use the cache
Some tests (like the incremental build in the nodejs image) depend on the fact
that a docker image is built from scratch and examines the build log.
With cache, the build log could contain different values than the build run
without the cache.
In case of the nodejs image, the result was that the incremental build
succeeded only every second time when run on the same machine.
Using no cache for docker build should make the tests more stable.