Harden the npm test
With 'set -o nounset' the NPM_REGISTRY usage made some issues, so make it more tolerant.
The test also expected that -testapp exists, which might not be true, changing to use 'sleep' call.
Finally, 'npm install' did not work for some reason with ubi8/s2i-base when run using 'bash -c'; with stdin it works for that image as well.