Add support for updating from tag from imagestream.
This commit adds a future to update imagestream
from our built image.
If version, which is updated is already present in JSON
file, then the ["from"]["name"] is updated  in JSON file.
If version does not exist yet, then
the script takes a first version, except latest,
and it uses it as a template for new version

Signed-off-by: Petr "Stone" Hracek <phracek@redhat.com>