Add ct_get_certificate_timestamp, ct_get_certificate_age_s, and ct_get_image_age_s functions
These functions are expected to be used for checking whether a certificate was just
generated during the container assembing (e.g age of the certificate is less than a minute).

The certificates generation on assembling the container with the application is important
for not having the container images with certificates being part of the image.