add function for checking the login status of oc 4
if not logged in, returns 1, sets TESTSUITE_RESULT to 1 and prints error
message to stdout

The function should be used in every container before running the
actual oc4 tests in order to make sure, that everything is setup right.
If this function returns non-zero status, the test script should
imediately exit, however on its own terms.