fix invalid docker build context in test suite
We have been symlinking ../ as common/ in tested container. However, the
container needs some files from common/ during build. As the common
symlink points to .., this was out of the reach for docker/podman
builder.

Implemented solution does not create symlink, but instead it packs
common/ into a tarball and use it in the tests.