adding sclorg tfaga wrapper
sclorg-tfaga-wrapper (Software Collections Organization's wrapper for
Testing Farm as a GitHub Action)

This wrapper tries to minimize lines of code needed to be written in
each
container repository when calling TFaGA.
This should lead to code reduction, dumplicity removal,
easier maintaining and adding features.

This action is only for Pull Request comment triggers.