use different trap functions for EXIT and SIGINT
so when the tests are stopped with SIGINT the final tasks are minimized
and the waiting time for results is reduced. That means, that image size
and stats are not shown in this case.

removing FAIL_QUICKLY option because if [ -n "${FAIL_QUICKLY}" ]
is evaluated as not true and is the last command of
the shell (which is almost in every container's tests) then the
shell ends with 1 and testsuite ends with FAILURE, as the return value
is caught and evaluated by the trap function.
Also we do not really use this feature as it is not fully implemented
for our needs.