OCP tests: examine the exit code in the trap function
in OCP3 tests we still use set -e for discovering test failure. However,
if the non-zero exit code is caught by trap function the whole bash ends
with exit code from the trap function itself and not with the exit code
with the tests.

This commit solves the issue by checking what value was returned from
the testsuite and exits the whole testsuite based on that.