Add .build-args-OS file to let specify build options
Podman used to set some capabilities by default, but does not any more
with version 4.4.0 and newer. These capabilities need to be set for
being able to build micro images that install RPMs into an alternative
path in the Dockerfile. Introducing .build-args-OS file means we will be
able to specify some arguments to docker build command for each
Dockerfile specifically, for example specify capabilities for the micro
images.