Onboarding OpenShift 4 MP+ cluster.
First of all we need to login into Cluster.
Credentials are part of TMT/FMF plan.

In order to turn on Shared cluster
SHARED_CLUSTER=true variable has to be set.

We need to be 100% sure in case of deleting
TenantNamespace. 'core-services-ocp--config' is not allowed.

All images built by our test suite has to be firstly pushed
to images.paas.redhat.com and then imported.
For now no other way is allowed.

Signed-off-by: Petr "Stone" Hracek <phracek@redhat.com>