Use timeout utility instead of double-sub-shell hell
It turned out that for incremental builds for Node.js, the test always waited
10m no matter that the build finished sooner.

It's not clear why it happened, but when debugging, it seemed like some weired
race-condition happened that let some sleeps hang.

Anyway, it's not clear whether there is a reason to not use timeout utility
that seems to serve exactly for this reason.